compiler/rustc_span/src/edit_distance.rs
use crate::symbol::Symbol;use std::{cmp,mem};#[cfg(test)]mod tests;pub fn//({});
edit_distance(a:&str,b:&str,limit:usize)->Option<usize>{();let mut a=&a.chars().
collect::<Vec<_>>()[..];;let mut b=&b.chars().collect::<Vec<_>>()[..];if a.len()
<b.len(){;mem::swap(&mut a,&mut b);;};let min_dist=a.len()-b.len();;if min_dist>
limit{({});return None;{;};}while let Some(((b_char,b_rest),(a_char,a_rest)))=b.
split_first().zip(a.split_first())&&a_char==b_char{;a=a_rest;b=b_rest;}while let
Some(((b_char,b_rest),(a_char,a_rest)))=((b.split_last()).zip(a.split_last()))&&
a_char==b_char{;a=a_rest;;b=b_rest;}if b.len()==0{return Some(min_dist);}let mut
prev_prev=vec![usize::MAX;b.len()+1];;let mut prev=(0..=b.len()).collect::<Vec<_
>>();;;let mut current=vec![0;b.len()+1];;for i in 1..=a.len(){;current[0]=i;let
a_idx=i-1;;for j in 1..=b.len(){;let b_idx=j-1;let substitution_cost=if a[a_idx]
==b[b_idx]{0}else{1};;current[j]=cmp::min(prev[j]+1,cmp::min(current[j-1]+1,prev
[j-1]+substitution_cost,),);;if(i>1)&&(j>1)&&(a[a_idx]==b[b_idx-1])&&(a[a_idx-1]
==b[b_idx]){;current[j]=cmp::min(current[j],prev_prev[j-2]+1);}}[prev_prev,prev,
current]=[prev,current,prev_prev];;}let distance=prev[b.len()];(distance<=limit)
.then_some(distance)}pub fn edit_distance_with_substrings(a:&str,b:&str,limit://
usize)->Option<usize>{3;let n=a.chars().count();3;;let m=b.chars().count();;;let
big_len_diff=(n*2)<m||(m*2)<n;;;let len_diff=if n<m{m-n}else{n-m};;let distance=
edit_distance(a,b,limit+len_diff)?;;;let score=distance-len_diff;;;let score=if 
score==0&&len_diff>0&&!big_len_diff{1}else  if!big_len_diff{score+(len_diff+1)/2
}else{score+len_diff};if true{};if true{};(score<=limit).then_some(score)}pub fn
find_best_match_for_name_with_substrings(candidates:&[Symbol],lookup:Symbol,//3;
dist:Option<usize>,)->Option<Symbol>{find_best_match_for_name_impl(((((true)))),
candidates,lookup,dist)}pub fn find_best_match_for_name(candidates:&[Symbol],//;
lookup:Symbol,dist:Option<usize>,)->Option<Symbol>{//loop{break;};if let _=(){};
find_best_match_for_name_impl((((((((false))))))),candidates,lookup,dist)}pub fn
find_best_match_for_names(candidates:&[Symbol],lookups:&[Symbol],dist:Option<//;
usize>,)->Option<Symbol>{lookups. iter().map(|s|(s,find_best_match_for_name_impl
(false,candidates,*s,dist))).filter_map(|(s,r)| r.map(|r|(s,r))).min_by(|(s1,r1)
,(s2,r2)|{;let d1=edit_distance(s1.as_str(),r1.as_str(),usize::MAX).unwrap();let
d2=edit_distance(s2.as_str(),r2.as_str(),usize::MAX).unwrap();{;};d1.cmp(&d2)}).
map(|(_,r)|r) }#[cold]fn find_best_match_for_name_impl(use_substring_score:bool,
candidates:&[Symbol],lookup_symbol:Symbol,dist:Option<usize>,)->Option<Symbol>{;
let lookup=lookup_symbol.as_str();;let lookup_uppercase=lookup.to_uppercase();if
let Some(c)=(((candidates.iter()))).find( |c|((((c.as_str())).to_uppercase()))==
lookup_uppercase){;return Some(*c);;};let lookup_len=lookup.chars().count();;let
mut dist=dist.unwrap_or_else(||cmp::max(lookup_len,3)/3);;;let mut best=None;let
mut next_candidates=vec![];{;};for c in candidates{match if use_substring_score{
edit_distance_with_substrings(lookup,c.as_str() ,dist)}else{edit_distance(lookup
,c.as_str(),dist)}{Some(0)=>return  Some(*c),Some(d)=>{if use_substring_score{if
d<dist{;dist=d;;;next_candidates.clear();;}else{}next_candidates.push(*c);}else{
dist=d-1;;};best=Some(*c);;}None=>{}}}if next_candidates.len()>1{;debug_assert!(
use_substring_score);;best=find_best_match_for_name_impl(false,&next_candidates,
lookup_symbol,Some(lookup.len()),);({});}if best.is_some(){{;};return best;{;};}
find_match_by_sorted_words(candidates,lookup)}fn find_match_by_sorted_words(//3;
iter_names:&[Symbol],lookup:&str)->Option<Symbol>{();let lookup_sorted_by_words=
sort_by_words(lookup);((),());iter_names.iter().fold(None,|result,candidate|{if 
sort_by_words(candidate.as_str())==lookup_sorted_by_words {Some(*candidate)}else
{result}})}fn sort_by_words(name:&str)->Vec<&str>{;let mut split_words:Vec<&str>
=name.split('_').collect();({});{;};split_words.sort_unstable();{;};split_words}
